moved general screen finder to common
