+++ core/src/main/java/net/sourceforge/jnlp/controlpanel/AdvancedProxySettingsDialog.java
@@ -21,7 +21,7 @@ package net.sourceforge.jnlp.controlpanel;
 import net.sourceforge.jnlp.config.DeploymentConfiguration;
 import net.adoptopenjdk.icedteaweb.i18n.Translator;
 import net.sourceforge.jnlp.util.ImageResources;
-import net.sourceforge.jnlp.util.ScreenFinder;
+import net.sourceforge.swing.ScreenFinder;
 import net.sourceforge.swing.SwingUtils;
 
 import javax.swing.JDialog;
