+++ core/src/main/java/net/sourceforge/jnlp/security/viewer/CertificateViewer.java
@@ -39,7 +39,7 @@ package net.sourceforge.jnlp.security.viewer;
 
 import net.sourceforge.jnlp.runtime.JNLPRuntime;
 import net.sourceforge.jnlp.util.ImageResources;
-import net.sourceforge.jnlp.util.ScreenFinder;
+import net.sourceforge.swing.ScreenFinder;
 import net.sourceforge.swing.SwingUtils;
 
 import javax.swing.JFrame;
