+++ core/src/main/java/net/sourceforge/jnlp/util/MD5SumWatcher.java
@@ -37,7 +37,6 @@ exception statement from your version.
 package net.sourceforge.jnlp.util;
 
 import net.adoptopenjdk.icedteaweb.IcedTeaWebConstants;
-import net.sourceforge.jnlp.util.logging.OutputController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
